+++ RefEngine/graphics/Prims.h
@@ -10,7 +10,7 @@ public:
 	Prims() = delete;
 
     static const std::vector<float> Quad_VerticesAndUVs;
-    static const std::vector<float> Quad_Indices;
+    static const std::vector<uint> Quad_Indices;
 
     static const std::vector<float> Triangle_Vertices;
     static const std::vector<float> Cube_Vertices;
